+++ next-auth.d.ts
@@ -1,11 +1,11 @@
-import { UserRole } from "@prisma/client"
+// import { UserRole } from "@prisma/client"
 
-export type ExtendedUser = DefaultSession["user"] & {
-  role: UserRole
-}
+// export type ExtendedUser = DefaultSession["user"] & {
+//   role: "ADMIN" | "USER"
+// }
 
-declare module "@auth/core" {
-  interface Session {
-    user: ExtendedUser
-  }
-}
+// declare module "next-auth" {
+//   interface Session {
+//     user: ExtendedUser
+//   }
+// }